Heretic: removed BLOCKMAP limit
Applied (with few corrections):
https://github.com/fabiangreffrath/crispy-doom/commit/423debf9c9ef18dfc794317a6104120adea39ee8

Fixes (partially) #30

+++ src/heretic/p_maputl.c
@@ -426,7 +426,7 @@ If the function returns false, exit with false without checking anything else.
 boolean P_BlockLinesIterator(int x, int y, boolean(*func) (line_t *))
 {
     int offset;
-    short *list;
+    long *list;
     line_t *ld;
 
     if (x < 0 || y < 0 || x >= bmapwidth || y >= bmapheight)
