+++ src/heretic/p_sight.c
@@ -94,7 +94,7 @@ boolean PTR_SightTraverse(intercept_t * in)
 boolean P_SightBlockLinesIterator(int x, int y)
 {
     int offset;
-    short *list;
+    long *list;
     line_t *ld;
     int s1, s2;
     divline_t dl;
